更新 Kconfig 文件以添加 RTL8822CU 的帮助信息，修改 Makefile 以支持新的构建配置，清理和优化多个源文件，删除不再使用的文件，更新 HAL 和 PHY 相关的结构以支持新的功能。

+++ include/rtw_rm.h
@@ -20,7 +20,7 @@ u8 rm_post_event_hdl(_adapter *padapter, u8 *pbuf);
 
 #define RM_TIMER_NUM 		32
 #define RM_ALL_MEAS		BIT(1)
-#define RM_ID_FOR_ALL(aid) (((u32)aid << 16) | RM_ALL_MEAS)
+#define RM_ID_FOR_ALL(aid)	((aid<<16)|RM_ALL_MEAS)
 
 #define RM_CAP_ARG(x) ((u8 *)(x))[4], ((u8 *)(x))[3], ((u8 *)(x))[2], ((u8 *)(x))[1], ((u8 *)(x))[0]
 #define RM_CAP_FMT "%02x %02x%02x %02x%02x"
